requirements.txt
# Feature.fm API Test Suite Dependencies
requests>=2.31.0
PyJWT>=2.8.0
colorama>=0.4.6
python-dotenv>=1.0.0

Flask==2.3.2
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0